+++ Changes.md
@@ -7,6 +7,7 @@
 
 - Fix broken Web UI response when using NewRelic and Rack 2.1.2+. [#4440]
 - Update APIs to use `UNLINK`, not `DEL`. [#4449]
+- Fix Ruby 2.7 warnings [#4412]
 
 6.0.4
 ---------